--- example/derivs_flat.txt
@@ -1,9 +1,0 @@
-CC(C)=CCC/C(C)=C/C=O
-CON=CC=C(C)CCC=C(C)C
-CC(C)=CCC\C(C)=C\C=O
-CCCCCC/C(=C\C1=CC=CC=C1)/C=O
-CCCCCC/C(C=NOC)=C\c1ccccc1
-CCCCCC/C(C=O)=C\c1ccccc1
-CON=CC1CC2C3C1C1(Cl)C(Cl)(Cl)C3(Cl)C3(Cl)C2C13Cl
-ClC12C3C4CC(C=O)C5C4C(Cl)(C13Cl)C(Cl)(Cl)C25Cl
-O=CC1CC2C3C1C1(Cl)C(Cl)(Cl)C3(Cl)C3(Cl)C2C13Cl
