--- example/derivs_struct.tsv
@@ -1,4 +1,0 @@
-orig	deriv. removed	deriv. added ...
-CC(C)=CCC\C(C)=C\C=O	CC(C)=CCC/C(C)=C/C=O	CC(C)=CCC/C(C)=C/C=O	CON=CC=C(C)CCC=C(C)C
-CCCCCC/C(=C\C1=CC=CC=C1)/C=O	CCCCCC/C(C=O)=C\c1ccccc1	CCCCCC/C(C=NOC)=C\c1ccccc1	CCCCCC/C(C=O)=C\c1ccccc1
-ClC12C3C4CC(C=O)C5C4C(Cl)(C13Cl)C(Cl)(Cl)C25Cl	O=CC1CC2C3C1C1(Cl)C(Cl)(Cl)C3(Cl)C3(Cl)C2C13Cl	CON=CC1CC2C3C1C1(Cl)C(Cl)(Cl)C3(Cl)C3(Cl)C2C13Cl	O=CC1CC2C3C1C1(Cl)C(Cl)(Cl)C3(Cl)C3(Cl)C2C13Cl
